fixed the abstract model and removed the extended classes
