+++ twister2/task/src/main/java/edu/iu/dsc/tws/task/batch/AbstractBatchSinkTask.java
@@ -9,11 +9,26 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-package edu.iu.dsc.tws.task.api;
+
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+package edu.iu.dsc.tws.task.batch;
 
 import edu.iu.dsc.tws.common.config.Config;
+import edu.iu.dsc.tws.task.api.IMessage;
+import edu.iu.dsc.tws.task.api.ISink;
+import edu.iu.dsc.tws.task.api.TaskContext;
 
-public class AbstractBatchSinkTask implements ISink {
+public abstract class AbstractBatchSinkTask implements ISink {
 
   private static final long serialVersionUID = -254264120110286748L;
   private TaskContext ctx;
@@ -26,6 +41,7 @@ public class AbstractBatchSinkTask implements ISink {
 
   @Override
   public void prepare(Config cfg, TaskContext context) {
-
+    this.ctx = context;
+    this.config = cfg;
   }
 }
